feat: connect plan permissions with premium features
